requirements.txt
alembic==1.8.1
bcrypt==3.2.2
fastapi==0.79.0
FastAPI-SQLAlchemy==0.2.1
uvicorn==0.18.2
psycopg2-binary==2.9.3
pydantic~=1.10.5
SQLAlchemy~=2.0.3
starlette~=0.19.1
pytz==2022.4
python-dateutil==2.8.2